DynamicFactorModeling.jl

Documentation for DynamicFactorModeling.jl

DynamicFactorModeling.DFMMeans — Type

DFMMeans(F::Array{Float64}, B::Array{Float64}, S::Array{Float64}, P::Array{Float64}, P2::Array{Float64})

Description: HDFM Bayesian estimator-generated latent factor and hyperparameter sample means (expected values).

Inputs:

F = MCMC-generated latent factor sample mean.

B = MCMC-generated observation equation regression coefficient sample means.

S = MCMC-generated observable variable idiosyncratic error disturbance variance sample means.

P = MCMC-generated latent factor autoregressive coefficient sample means.

P2 = MCMC-generated idiosyncratic error autoregressive coefficient sample means.

DynamicFactorModeling.DFMResults — Type

DFMResults(F::Array{Float64}, B::Array{Float64}, S::Array{Float64}, P::Array{Float64}, P2::Array{Float64}, means::DFMMeans)

Description: HDMF Bayesian estimator-generated MCMC posterior distribution samples and their means for latent factors and hyperparameters.

Inputs:

F = MCMC-generated latent factor sample.

B = MCMC-generated observation equation regression coefficient sample.

S = MCMC-generated observable variable idiosyncratic error disturbance variance sample.

P = MCMC-generated latent factor autoregressive coefficient sample.

P2 = MCMC-generated idiosyncratic error autoregressive coefficient sample.

means = HDFM Bayesian estimator-generated latent factor and hyperparameter sample means (expected values).

DynamicFactorModeling.DFMStruct — Type

DFMStruct(factorlags::Int64, errorlags::Int64, ndraws::Int64, burnin::Int64)

Description: 1-level DFM lag structure specification and MCMC sample size for Bayesian estimation.

Inputs:

factorlags = Number of lags in the autoregressive specification of the latent factors.

errorlags = Number of lags in the autoregressive specification of the observable variable idiosyncratic errors.

ndraws = Number of MCMC draws used for posterior distributions.

burnin = Number of initial MCMC draws discarded.

DynamicFactorModeling.HDFM — Type

HDFM(nlevels, nvar, nfactors, fassign, flags, varlags, varcoefs, fcoefs, fvars, varvars)

Description: Creates an object of type HDFM that contains all parameters necessary to specify a multi-level linear dynamic factor data-generating process. This is a convenient alternative to specifying an HDFM directly in state-space form.

Inputs:

nlevels = number of levels in the multi-level model structure.

nvar = number of variables.

nfactors = number of factors for each level (vector of length nlevels).

fassign = determines which factor is assigned to which variable for each level (integer matrix of size nvar × nlevels).

flags = number of autoregressive lags for factors of each level (factors of the same level are restricted to having the same number of lags; vector of length nlevels).

varlags = number of observed variable error autoregressive lags (vector of length nvar).

varcoefs = vector of coefficients for each variable in the observation equation (length 1+nlevels, where first entry represents the intercept).

fcoefs = list of nlevels number of matrices, for which each row contains vectors of the autoregressive lag coefficients of the corresponding factor.

fvars = list of nlevels number of vectors, where each entry contains the disturbance variance of the corresponding factors.

varvars = vector of nvar number of entries, where each entry contains the innovation variance of the corresponding variable.

DynamicFactorModeling.HDFMStruct — Type

HDFMStruct(nlevels::Int64, nfactors::Array{Int64,1}, factorassign::Array{Int64,2}, factorlags::Array{Int64,1}, errorlags::Array{Int64,1}, ndraws::Int64, burnin::Int64)

Description: Multi-level/hierarchical DFM (HDFM) level, factor assignment, and lag structure specification, and MCMC sample size for Bayesian estimation.

Inputs:

nlevels = Number of levels in the HDFM specification.

nvars = Number of observable variables in the HDFM specification.

nfactors = Number of factor per level in the HDFM specification.

factorassign = Factors assigned to each variable across all levels.

factorlags = Number of lags in the autoregressive specification of the latent factors.

errorlags = Number of lags in the autoregressive specification of the observable variable idiosyncratic errors.

ndraws = Number of MCMC draws used for posterior distributions.

burnin = Number of initial MCMC draws discarded.

DynamicFactorModeling.SSModel — Type

SSModel(H, A, F, μ, R, Q, Z)

Description: A type object containing all parameters necessary to specify a data-generating process in state-space form. Measurement Equation: y{t} = H β{t} + A z{t} + e{t} Transition Equation: β{t} = μ + F β{t-1} + v{t} e{t} ~ i.i.d.N(0,R) v{t} ~ i.i.d.N(0,Q) z{t} ~ i.i.d.N(0,Z) E(et vs') = 0

Inputs:

H = measurement equation state vector coefficient matrix.

A = measurement equation predetermined vector coefficient matrix.

F = state equation companion matrix.

μ = state equation intercept vector.

R = measurement equation error covariance matrix.

Q = state equation innovation covariance matrix.

Z = pretermined vector covariance matrix.

DynamicFactorModeling.KN1LevelEstimator — Method

KNSingleFactorEstimator(data::Array{Float64,2}, dfm::DFMStruct)

Description: Estimate a single-factor DFM using the Kim-Nelson approach.

Inputs:

data = Matrix with each column being a data series.

dfm = Model structure specification.

Outputs:

results = HDMF Bayesian estimator-generated MCMC posterior distribution samples and their means for latent factors and hyperparameters.

DynamicFactorModeling.KN2LevelEstimator — Method

KN2LevelEstimator(data::Array{Float64,2}, hdfm::HDFMStruct)

Description: Estimate a two-level HDFM using the Kim-Nelson approach. Both the latent factors and hyperparameters are estimated using the Bayesian approach outlined in Kim and Nelson (1999).

Inputs:

data = Matrix with each column being a data series.

hdfm = Model structure specification.

Outputs:

results = HDMF Bayesian estimator-generated MCMC posterior distribution samples and their means for latent factors and hyperparameters.

DynamicFactorModeling.KNFactorSampler — Method

KNFactorSampler(data_y, ssmodel)

Description: Draw a sample series of dynamic factor from conditional distribution in Ch 8, Kim & Nelson (1999). Measurement Equation: y{t} = H{t} β{t} + A z{t} + e{t}. Transition Equation: β{t} = μ + F β{t-1} + v{t}; e{t} ~ i.i.d.N(0,R); v{t} ~ i.i.d.N(0,Q); z{t} ~ i.i.d.N(0,Z); E(et v_s') = 0.

Inputs:

data = observed data

H = measurement eq. state coef. matrix

A = measurement eq. exogenous coef. matrix

F = state eq. companion matrix

μ = state eq. intercept term

R = covariance matrix on measurement disturbance

Q = covariance matrix on state disturbance

Z = covariance matrix on predetermined var vector

DynamicFactorModeling.OW1LevelEstimator — Method

OW1LevelEstimator(data::Array{Float64,2}, dfm::DFMStruct)

Description: Estimate a single-factor DFM using the Otrok-Whiteman approach.

Inputs:

data = Matrix with each column being a data series.

dfm = Model structure specification.

Outputs:

results = HDMF Bayesian estimator-generated MCMC posterior distribution samples and their means for latent factors and hyperparameters.

DynamicFactorModeling.OW2LevelEstimator — Method

OW2LevelEstimator(data::Array{Float64,2}, hdfm::HDFMStruct)

Description: Estimate a two-level HDFM using the Kim-Nelson approach. Both the latent factors and hyperparameters are estimated using the Bayesian approach outlined in Kim and Nelson (1999).

Inputs:

data = Matrix with each column being a data series.

hdfm = Model structure specification.

Outputs:

results = HDMF Bayesian estimator-generated MCMC posterior distribution samples and their means for latent factors and hyperparameters.

DynamicFactorModeling.PCA1LevelEstimator — Method

PCA1LevelEstimator(data::Array{Float64,2}, dfm::DFMStruct)

Description: Estimate a single-factor DFM using the Kim-Nelson approach.

Inputs:

data = Matrix with each column being a data series.

dfm = Model structure specification.

Outputs:

results = HDMF Bayesian estimator-generated MCMC posterior distribution samples and their means for latent factors and hyperparameters.

DynamicFactorModeling.PCA2LevelEstimator — Method

PCA2LevelEstimator(data::Array{Float64,2}, hdfm::HDFMStruct)

Description: Estimate a two-level HDFM using the PCA approach. The latent factors are estimated using PCA, while the hyperparameters are estimated using the Bayesian approach outlined in Kim and Nelson (1999).

Inputs:

data = Matrix with each column being a data series.

hdfm = Model structure specification.

Outputs:

results = HDMF Bayesian estimator-generated MCMC posterior distribution samples and their means for latent factors and hyperparameters.

DynamicFactorModeling.ar — Method

ar(y, x, p, b0_, B0__, r0_, R0__, v0_, d0_, b0, s20, phi0, xvar, nfc, facts, capt, nreg, Size)

Description: Generate βi , ϕi, and σ^2_i estimates based on the posterior distributions (eqs. 6-8) on pg. 1002 of Otrok and Whiteman (1998).

Inputs:

y = observed dependent variable / data.

x = measurement equation regressors (intercept + factor).

p = number of state eq. lags.

b0_ = prior mean of factor loading (factor coefficient in measurement equation).

B0__ = prior variance of factor loading (factor coefficient in measurement equation).

r0_ = prior mean of phi (idiosyncratic AR polynomial).

R0__ = prior precision of phi (idiosyncratic AR polynomial).

v0_ = inverted gamma parameter of innovation variances.

d0_ = inverted gamma parameter of innovation variances.

b0 = old draw of factor loading.

s20 = old draw of innovation variances.

phi0 = old draw of phi (idiosyncratic AR polynomial).

xvar = measurement eq. regressor index.

nfc = number of factors.

facts = factor draw.

capt = total number of time periods / observations in the sample.

nreg = number of regressors in each observable equation, constant plus K factors

Size = number of variables each factor loads on

Outputs:

b0 = β coefficient hyperparameter estimates.

s20 = σ^2 coefficient hyperparameter estimate.

phi1 = ϕ coefficient hyperparameter estimates.

facts = factor estimate.

DynamicFactorModeling.ar_LJ — Method

DynamicFactorModeling.arfac — Method

DynamicFactorModeling.convertHDFMtoSS — Method

convertHDFMtoSS(hdfm)

Description: Converts an HDFM object to an SSModel object.

Inputs:

hdfm = HDFM object containing all HDFM DGP parameters.

Outputs:

ssmodel = SSModel object containing all state-space model parameters.

DynamicFactorModeling.createSSforHDFM — Method

createSSforHDFM(hdfm)

Description: Create state-space form coefficient and variance matrices for an HDFM object.

Inputs:

hdfm = HDFM object with all parameters necessary to specify a multi-level linear dynamic factor data-generating process.

DynamicFactorModeling.diagrv — Method

diagrv(X,V)

Description: Replace the diagonal of matrix X with vector V.

Inputs:

X = matrix of interest

V = new diagonal of X

Outputs:

X_new = X with new diagonal V

DynamicFactorModeling.firstComponentFactor — Method

DynamicFactorModeling.gendiff — Method

gendiff(z, phi)

Description: Difference a series z using the lag coefficients provided in phi.

Inputs:

z = time series.

phi = lag coefficient vector.

Outputs:

zgdiff = quasi-differenced version of the inputted series z.

DynamicFactorModeling.ident — Method

ident(p)

Description: Create dense identity matrix of dimension p × p.

Inputs:

p = number of columns and rows of identity matrix.

Outputs:

ident_mat = p × p dense identity matrix with Float64.

DynamicFactorModeling.invpd — Method

invpd(X)

Description: Invert matrix X using augmented eigenvalues if X is not positive-definite.

Inputs:

X = matrix of interest

Output:

X_inv = inverse of input matrix X

DynamicFactorModeling.kalmanFilter — Method

kalmanFilter(data, ssmodel)

Description: Apply Kalman filter to observed data. Measurement Equation: y{t} = H{t} β{t} + A z{t} + e{t} . Transition Equation: β{t} = μ + F β{t-1} + v{t}; e{t} ~ i.i.d.N(0,R); v{t} ~ i.i.d.N(0,Q); z{t} ~ i.i.d.N(0,Z); E(et v_s') = 0.

Inputs:

data = observed data

H = measurement eq. state coef. matrix

A = measurement eq. exogenous coef. matrix

F = state eq. companion matrix

μ = state eq. intercept term

R = covariance matrix on measurement disturbance

Q = covariance matrix on state disturbance

Z = covariance matrix on predetermined var vector

DynamicFactorModeling.kalmanSmoother — Method

kalmanSmoother(data, ssmodel)

Description: Apply Kalman smoother to observed data. Measurement Equation: y{t} = H{t} β{t} + A z{t} + e{t}. Transition Equation: β{t} = μ + F β{t-1} + v{t}; e{t} ~ i.i.d.N(0,R); v{t} ~ i.i.d.N(0,Q); z{t} ~ i.i.d.N(0,Z); E(et v_s') = 0.

Inputs:

data = observed data

H = measurement eq. state coef. matrix

A = measurement eq. exogenous coef. matrix

F = state eq. companion matrix

μ = state eq. intercept term

R = covariance matrix on measurement disturbance

Q = covariance matrix on state disturbance

Z = covariance matrix on predetermined var vector

DynamicFactorModeling.linearRegressionSampler — Method

linearRegressionSampler(Y,X)

Description: Estimate β and σ^2 in Y = Xβ + e, e ~ N(0,σ^2 I_T). Generate samples of β and σ^2.

Inputs:

Y = Dependent data matrix

X = Independent data matrix

DynamicFactorModeling.linearRegressionSamplerRestrictedVariance — Method

linearRegressionSamplerRestrictedVariance(Y, X, σ2)

Description: Estimate β and σ^2 in Y = Xβ + e, e ~ N(0,σ^2 I_T), where σ^2 is restricted to some chosen value. Generate samples of β and σ^2.

Inputs:

Y = Dependent data matrix

X = Independent data matrix

σ2 = Restricted error variance

DynamicFactorModeling.rnchisq — Method

rnchisq(m)

Description: Draw sample from a chi-square distribution.

Inputs:

m = degrees of freedom.

Outputs:

g = draw from a chi-square distribution with m degrees of freedom.

DynamicFactorModeling.seqa — Method

seqa(a, b, c)

Description: Create a sequence with c # of inputs, starting at a with b increments.

Inputs:

a = starting value of the sequence.

b = increment size.

c = number of increments.

Outputs:

seq = sequence with c # of inputs, starting at a with b increments.

DynamicFactorModeling.seqm — Method

seqm(a,b,c)

Description: Prouce a sequence of values.

Inputs:

a = initial value of sequence

b = increment scaling size

c = number of values in the sequence

Outputs:

seq = geometric sequence

DynamicFactorModeling.sigbig — Method

sigbig(phi, p, capt)

Description: Create T×T matrix S^(-1) (pg. 1003 of Otrok and Whiteman (1998)).

Inputs:

phi = Lag coefficients associated with idiosyncratic error autoregression.

p = Number of lags.

capt = total number of time periods in the sample.

Outputs:

Si = T×T matrix S^(-1).

DynamicFactorModeling.sigmat — Method

sigmat(phi, p)

Description: Yield the stationary covariance matrix of the first p errors (pg 1001 of Otrok and Whiteman (1998)).

Inputs:

phi = Lag coefficients associated with idiosyncratic error autoregression.

p = Number of lags.

Outputs:

Stationary covariance matrix of the first p errors.

DynamicFactorModeling.sim_MvNormal — Method

DynamicFactorModeling.sim_MvNormal_alt — Method

DynamicFactorModeling.simulateSSModel — Method

simulateSSModel(num_obs, ssmodel::SSModel)

Description: Generate data from a DGP in state space form. Measurement Equation: y{t} = H β{t} + A z{t} + e{t} Transition Equation: β{t} = μ + F β{t-1} + v{t} e{t} ~ i.i.d.N(0,R) v{t} ~ i.i.d.N(0,Q) z{t} ~ i.i.d.N(0,Z) E(et vs') = 0

Inputs:

num_obs = number of observations

ssmodel = SSModel object containing all parameters necessary to specify a data-generating process in state-space form.

DynamicFactorModeling.transp_dbl — Method

transp_dbl(A)

Description: Transpose matrix, so that the output type is Array{Float64,2}.

Inputs:

A = matrix.

Outputs:

transp_A = dense transposed matrix A.

DynamicFactorModeling.transp_int — Method

transp_int(A)

Description: Transpose matrix, so that the output type is Array{Int64,2}.

Inputs:

A = matrix.

Outputs:

transp_A = dense transposed matrix A.

DynamicFactorModeling.vardecomp2level — Method

vardecomp2level(data::Array{Float64, 2}, factor::Array{Float64}, betas::Array{Float64}, factorassign::Array{Float64})

Description: Compute the portion of the variation of each observable series that may be attributed to their corresponding/assigned latent factors across all levels.

Inputs:

data = Matrix with each column representing a data series.

factor = Matrix containing latent factor estimates.

betas = Matrix containing observation equation coefficient parameter estimates.

factorassign = Matrix containing the indeces of factors across all levels (columns) assigned to each observable series (rows).

Output:

vardecomps = Matrix containing the variance contributions of factors across all levels (columns) corresponding to each observable series (rows).